+++ Tests/MapboxNavigationTests/NavigationCameraTests.swift
@@ -461,11 +461,7 @@ class NavigationCameraTests: TestCase {
         let expectedCenterCoordinate = CLLocationCoordinate2D(latitude: 0.0, longitude: 0.0)
         navigationViewportDataSource.followingMobileCamera.center = expectedCenterCoordinate
         
-        
-        // Since zoomUpdatesAllowed is disabled it is expected that `CameraOptions.zoom`, which was
-        // returned from the `ViewportDataSourceDelegateMock` will be the median value of the zoom range.
-        let zoomRange = navigationViewportDataSource.options.followingCameraOptions.zoomRange
-        let expectedZoom = CGFloat((zoomRange.lowerBound + zoomRange.upperBound) / 2)
+        let expectedZoom: CGFloat = 11.1
         navigationViewportDataSource.followingMobileCamera.zoom = expectedZoom
         
         let expectedBearing = 22.2